feat(#398) - improve logic of deactivation the last user with root role.

+++ application/app/Console/Commands/DetachRolesFromDeactivatedUsers.php
@@ -7,6 +7,7 @@ use App\Util\DateUtil;
 use Exception;
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Console\Isolatable;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Facades\Date;
 use Illuminate\Support\Facades\DB;
 use Throwable;
@@ -30,6 +31,13 @@ class DetachRolesFromDeactivatedUsers extends Command implements Isolatable
                         ->select('id')
                         ->whereDate('deactivation_date', '<=', Date::now(DateUtil::ESTONIAN_TIMEZONE)->format('Y-m-d'))
                 )->each(function (InstitutionUserRole $institutionUserRole) {
+                    if ($institutionUserRole->role->is_root && !$this->institutionHasAnotherUserWithRootRole($institutionUserRole)) {
+                        $institutionUser = $institutionUserRole->institutionUser;
+                        $institutionUser->deactivation_date = null;
+                        $institutionUser->save();
+                        return true;
+                    }
+
                     try {
                         $institutionUserRole->delete();
                     } catch (Exception) {
@@ -40,5 +48,18 @@ class DetachRolesFromDeactivatedUsers extends Command implements Isolatable
 
             // TODO: audit log
         });
+    }
+
+    private function institutionHasAnotherUserWithRootRole(InstitutionUserRole $institutionUserRole): bool
+    {
+        $institutionUser = $institutionUserRole->institutionUser;
+        return InstitutionUserRole::query()
+            ->whereRelation('role', function (Builder $query) use ($institutionUser) {
+                $query->where('is_root', true)
+                    ->where('institution_id', $institutionUser->institution_id);
+            })->whereRelation('institutionUser', function (Builder $query) use ($institutionUser) {
+                $query->where('institution_id', $institutionUser->institution_id)
+                    ->whereNot('id', $institutionUser->id);
+            })->exists();
     }
 }
